refactor: Simplify "add_param_section_to_docstring"
This change replaces the if statement block used to determine the
position of the parameter name and type with conditional expressions
for better conciseness and readability. The remaining if statements
for inserting the parameter title and removing the last line in
docstrings with 2 base sections were refactored into single-line
expressions.

Additionally, the following variables were renamed:
- "Pos_name" -> "name_pos"
- "Pos_type" -> "type_pos"