refactor(insert_docs): Improve insertion position detection
Previously, Codedocs inserted the docstring at the cursor's position.
This behavior worked well if the goal was for users to place their
cursor directly on the structure to document. However, it limited
flexibility, as users couldn't generate a docstring from any position
within the function.

This change modifies the insertion logic to use the position of the
node to be documented in the buffer, rather than the cursor's position.
This makes the plugin more automatic, flexible, and less prone to errors.