feat: Add basic class docs generation and queries module
This update implements core logic to parse class nodes and extract
relevant information such as attributes.

It introduces the queries module to leverage Treesitter queries for
a more declarative and maintainable approach to node traversal. This
improves code clarity compared to the for-loop traversal currently
used for function parsing.

Class docstring generation is currently supported for the following
languages:

- Python
- Java
- Kotlin
- TypeScript
- JavaScript