refactor(styles): Split into individual dirs/modules
Splits language style files into individual directories and modules,
since having all styles for all structures in a language in a single
module made it harder to debug, update, and even read the options due to
the nesting.

Additionally, the manager module is now in charge of retrieving the
appropriate style, leaving the builder module fully focused on building
the docstring.